% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_band.R, R/fregion.band.R
\name{band}
\alias{band}
\alias{confidence_band}
\title{Makes confidence bands}
\usage{
confidence_band(x, cov, tau = NULL, t0 = NULL, N = 1,
  type = c("FFSCB.z", "FFSCB.t", "KR.z", "KR.t", "BEc", "Bs", "naive.t"),
  conf.level = c(0.95), grid.size = 200, Bs.sim.size = 10000,
  n_int = 10, tol = NULL)

confidence_band(x, cov, tau = NULL, t0 = NULL, N = 1,
  type = c("FFSCB.z", "FFSCB.t", "KR.z", "KR.t", "BEc", "Bs", "naive.t"),
  conf.level = c(0.95), grid.size = 200, Bs.sim.size = 10000,
  n_int = 10, tol = NULL)
}
\arguments{
\item{x}{Functional parameter estimate. It can be either a vector or \link{fd} object from \link{fda}.}

\item{cov}{N * Cov(X), in which X is the functional estimator. It can be either matrix or \link{bifd} object from \link{fda}. The eigen decomposition of Cov(X) can be used instead.}

\item{tau}{Pointwise standard deviation of the standardized and differentiated sample functions. Can be estimated by tau_fun().}

\item{t0}{Parameter t0 of the fast and fair simultaneous confidence bands.}

\item{N}{It should be '1' if 'cov' is the covariance operator for X itself, which is the default value.}

\item{type}{The band(s) to be constructed.
\itemize{
  \item FFSCB.z : Fast'n'Fair (adaptive) simultaneous confidence band based for a Gaussian functional parameter estimate.
  \item FFSCB.t : Fast'n'Fair (adaptive) simultaneous confidence band based for a t-distributed functional parameter estimate.
  \item KR.z : The constant simultaneous confidence band based on the classical Kac-Rice (KR) formula for Gaussian random functions.
  \item KR.t : The constant simultaneous confidence band based on the classical Kac-Rice (KR) formula for t-distributed random functions.
  \item BEc : The suggested modified Scheffe style band from hyper-ellipsoie Ec, which uses up to the very last dimension.
  \item Bs : Parametric bootstrap simultaneous confidence band, similar to the one appeard in Degras(2011) (for comparision purpose)
  \item naive.t : A collection of point-wise t-intervals. (for comparision purpose)
}}

\item{conf.level}{A vector of confidence levels for the bands to achieve.}

\item{grid.size}{This determines on how fine grid the bands will be constructed before converted as an `fd' object. This parameter is used only when 'x' is fd object and 'cov' is bifd object.}

\item{Bs.sim.size}{This determines bootstrap sample size for Bs}

\item{n_int}{Number of intervals for the piecewise linear confidence bounds.}

\item{tol}{Controls the tolerance value used by stats::optimize(). The default (tol=NULL) leads to the functions' default values.}

\item{x}{Functional parameter estimate. It can be either a vector or \link{fd} object from \link{fda}.}

\item{cov}{N * Cov(X), in which X is the functional estimator. It can be either matrix or \link{bifd} object from \link{fda}. The eigen decomposition of Cov(X) can be used instead.}

\item{tau}{Pointwise standard deviation of the standardized and differentiated sample functions. Can be estimated by tau_fun().}

\item{t0}{Parameter t0 of the fast and fair simultaneous confidence bands.}

\item{N}{It should be '1' if 'cov' is the covariance operator for X itself, which is the default value.}

\item{type}{The band(s) to be constructed.
\itemize{
  \item FFSCB.z : Fast'n'Fair (adaptive) simultaneous confidence band based for a Gaussian functional parameter estimate.
  \item FFSCB.t : Fast'n'Fair (adaptive) simultaneous confidence band based for a t-distributed functional parameter estimate.
  \item KR.z : The constant simultaneous confidence band based on the classical Kac-Rice (KR) formula for Gaussian random functions.
  \item KR.t : The constant simultaneous confidence band based on the classical Kac-Rice (KR) formula for t-distributed random functions.
  \item BEc : The suggested modified Scheffe style band from hyper-ellipsoie Ec, which uses up to the very last dimension.
  \item Bs : Parametric bootstrap simultaneous confidence band, similar to the one appeard in Degras(2011) (for comparision purpose)
  \item naive.t : A collection of point-wise t-intervals. (for comparision purpose)
}}

\item{conf.level}{A vector of confidence levels for the bands to achieve.}

\item{grid.size}{This determines on how fine grid the bands will be constructed before converted as an `fd' object. This parameter is used only when 'x' is fd object and 'cov' is bifd object.}

\item{Bs.sim.size}{This determines bootstrap sample size for Bs}

\item{n_int}{Number of intervals for the piecewise linear confidence bounds.}

\item{tol}{Controls the tolerance value used by stats::optimize(). The default (tol=NULL) leads to the functions' default values.}
}
\value{
confidence_band Either a collection of vector valued bands or `fd' object whose objectname is changed to confidence_band.

confidence_band Either a collection of vector valued bands or `fd' object whose objectname is changed to confidence_band.
}
\description{
Makes confidence bands

Makes confidence bands
}
\examples{
# Generate a sample
p <- 200 ; N <- 80 ; rangeval = c(0,1)
grid  <- make_grid(p, rangevals=rangeval)
mu0   <- meanf_poly(grid,c(0,1)) ; names(mu0) = grid
mu    <- meanf_poly(grid,c(0,1.1)) ; names(mu) = grid
cov.m <- make_cov_m(cov.f = covf.st.matern, grid=grid, cov.f.params=c(2/2,1,1))
x     <- make_sample(mu,cov.m,N)

# Find the estimate and covariance
hat.mu       <- rowMeans(x)
hat.cov.m    <- crossprod(t(x - hat.mu)) / (N-1)
hat.tau.v    <- tau_fun(x)

# Make and visualize/compare confidence bands
b <- confidence_band(x=hat.mu,cov=hat.cov.m,tau=hat.tau.v,N=N,
                  type=c("FFSCB.t", "Bs","BEc","naive.t"),conf.level=c(0.95))
plot(b)
# Generate a sample
p <- 200 ; N <- 80 ; rangeval = c(0,1)
grid  <- make_grid(p, rangevals=rangeval)
mu0   <- meanf_poly(grid,c(0,1)) ; names(mu0) = grid
mu    <- meanf_poly(grid,c(0,1.1)) ; names(mu) = grid
cov.m <- make_cov_m(cov.f = covf.st.matern, grid=grid, cov.f.params=c(2/2,1,1))
x     <- make_sample(mu,cov.m,N)

# Find the estimate and covariance
hat.mu       <- rowMeans(x)
hat.cov.m    <- crossprod(t(x - hat.mu)) / (N-1)
hat.tau.v    <- tau_fun(x)

# Make and visualize/compare confidence bands
b <- confidence_band(x=hat.mu,cov=hat.cov.m,tau=hat.tau.v,N=N,
                  type=c("FFSCB.t", "Bs","BEc","naive.t"),conf.level=c(0.95))
plot(b)
}
\references{
\itemize{
   \item Liebl, D. and Reimherr, M. (2019). Fast and fair simultaneous confidence bands.
   \item  Choi, H. and Reimherr, M. (2018). A geometric approach to confidence regions and bands for functional parameters. Journal of the Royal Statistical Society: Series B (Statistical Methodology) 80 239-260.
}

\itemize{
   \item Liebl, D. and Reimherr, M. (2019). Fast and fair simultaneous confidence bands.
   \item  Choi, H. and Reimherr, M. (2018). A geometric approach to confidence regions and bands for functional parameters. Journal of the Royal Statistical Society: Series B (Statistical Methodology) 80 239-260.
}
}
