% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Biomechanics}
\alias{Biomechanics}
\title{Biomechanics Data (Biomechanics_Data)}
\format{Variables:
\describe{
  \item{Stance_Phase}{in percent}
  \item{Extra_Cushioned_1:Extra_Cushioned_11}{Torque curves (N m / kg) when running with extra cushioned running shoes.}
  \item{Normal_Cushioned_1:Normal_Cushioned_11}{Torque curves (N m / kg) when running with normal cushioned running shoes.}
}}
\usage{
Biomechanics
}
\description{
Paired sample of n=11 subjects.
}
\details{
Each subject performed one running stride with an extra cushioned 
running shoe and one running stride with a normally cushioned 
running shoe.

Measured outcome: 
Torque (N m / kg) at ankle joint in sagittal pane during the stance phase
standardized with respect to the bodyweight of the subjects.
}
\keyword{datasets}
