% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/confidence_band.R
\name{confidence_band}
\alias{confidence_band}
\title{Makes confidence bands}
\usage{
confidence_band(x, cov.x, tau = NULL, t0 = NULL, df = NULL,
  type = c("FFSCB.z", "FFSCB.t", "KR.z", "KR.t", "BEc", "Bs", "naive.t"),
  conf.level = 0.95, grid.size = 200, Bs.sim.size = 10000,
  n_int = 10)
}
\arguments{
\item{x}{Functional parameter estimate (for instance, the empirical mean function). It can be either a vector or \link{fd} object from \link{fda}.}

\item{cov.x}{Cov(x), in which x is the functional estimator (for instance, the covariance function of the empirical mean function). It can be either matrix or \link{bifd} object from \link{fda}. The eigen decomposition of Cov(X) can be used instead.}

\item{tau}{Pointwise standard deviation of the standardized and differentiated sample functions. Can be estimated by tau_fun().}

\item{t0}{Parameter t0 of the fast and fair simultaneous confidence bands.}

\item{df}{Degrees of freedom parameter for the t-distribution based bands 'FFSCB.t', 'KR.t', and 'naive.t'. If x is the empirical mean function, set df=n-1, where n denotes the sample size.}

\item{type}{The band(s) to be constructed.
\itemize{
  \item FFSCB.z : Fast'n'Fair (adaptive) simultaneous confidence band based for a Gaussian functional parameter estimate.
  \item FFSCB.t : Fast'n'Fair (adaptive) simultaneous confidence band based for a t-distributed functional parameter estimate.
  \item minW.z : 
  \item KR.z : The constant simultaneous confidence band based on the classical Kac-Rice (KR) formula for Gaussian random functions.
  \item KR.t : The constant simultaneous confidence band based on the classical Kac-Rice (KR) formula for t-distributed random functions.
  \item BEc : The suggested modified Scheffe style band from hyper-ellipsoie Ec, which uses up to the very last dimension.
  \item Bs : Parametric bootstrap simultaneous confidence band, similar to the one appeard in Degras(2011) (for comparision purpose)
  \item naive.t : A collection of point-wise t-intervals. (for comparision purpose)
}}

\item{conf.level}{A vector of confidence levels for the bands to achieve.}

\item{grid.size}{This determines on how fine grid the bands will be constructed before converted as an `fd' object. This parameter is used only when 'x' is fd object and 'cov.x' is bifd object.}

\item{Bs.sim.size}{This determines bootstrap sample size for Bs}

\item{n_int}{Number of intervals for the piecewise linear confidence bounds.}
}
\value{
confidence_band Either a collection of vector valued bands or `fd' object whose objectname is changed to confidence_band.
}
\description{
Makes confidence bands
}
\examples{
# Generate a sample
p          <- 200 
N          <- 80 
grid       <- make_grid(p, rangevals=c(0,1))
mu0        <- meanf_poly(grid,c(0,1))   
names(mu0) <- grid
mu         <- meanf_poly(grid,c(0,1.1)) 
names(mu)  <- grid
cov.m      <- make_cov_m(cov.f = covf.st.matern, grid=grid, cov.f.params=c(2/2,1))
sample     <- make_sample(mu,cov.m,N)

# Compute the estimate and its covariance
hat.mu     <- rowMeans(sample)
hat.cov    <- crossprod(t(sample - hat.mu)) / N
hat.cov.mu <- hat.cov / N

# Compute the tau-parameter (for the KR- and FFSCB-bands)
hat.tau    <- tau_fun(sample)

# Make and plot confidence bands
b <- confidence_band(x=hat.mu, cov.x=hat.cov.mu, tau=hat.tau, df=N-1,
                     type=c("FFSCB.t", "Bs","BEc","naive.t"),
                     conf.level  = 0.95)
plot(b)
lines(x=grid, y=mu0, lty=2)
}
\references{
\itemize{
   \item Liebl, D. and Reimherr, M. (2019). Fast and fair simultaneous confidence bands.
   \item  Choi, H. and Reimherr, M. (2018). A geometric approach to confidence regions and bands for functional parameters. Journal of the Royal Statistical Society: Series B (Statistical Methodology) 80 239-260.
}
}
