% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{cov_fragments}
\alias{cov_fragments}
\title{Estimate covariance function from fragmentary functional data}
\usage{
cov_fragments(X_mat)
}
\arguments{
\item{X_mat}{nxp Matrix with fragmentary observed functional data of subjects i=1,...,p stored column-wise. Missing data are assumed to be set to NA.}
}
\description{
Estimate covariance function from fragmentary functional data
}
