% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covf.R
\name{covf.nonst.matern}
\alias{covf.nonst.matern}
\title{Modified Matern Covariance Function (varying roughness parameter)}
\usage{
covf.nonst.matern(x1, x2, params = c(3/2, 1/2, 1))
}
\arguments{
\item{x1}{First argument of cov(x1,x2).}

\item{x2}{Second argument of cov(x1,x2).}

\item{params}{Covariance function parameters: params=c(nu1, nu2, sigma).}
}
\description{
Modified Matern Covariance Function (varying roughness parameter)
}
