% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covf.R
\name{covf.st.matern}
\alias{covf.st.matern}
\title{Matern Covariance Function}
\usage{
covf.st.matern(x1, x2, params = c(1/2, 1, 1))
}
\arguments{
\item{x1}{First argument of cov(x1,x2).}

\item{x2}{Second argument of cov(x1,x2).}

\item{params}{Matern covariance function parameters: params=c(nu, l, sigma).}
}
\description{
Matern Covariance Function
}
