% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/covf.R
\name{covf.st.matern.warp.logit_inv}
\alias{covf.st.matern.warp.logit_inv}
\title{Matern Covariance Function with inverse logit warped time-scale}
\usage{
covf.st.matern.warp.logit_inv(x1, x2, params = c(1/2, 1, 1))
}
\arguments{
\item{x1}{First argument of cov(x1,x2).}

\item{x2}{Second argument of cov(x1,x2).}

\item{params}{Matern covariance function parameters: params=c(nu, l, sigma).}
}
\description{
Matern Covariance Function with inverse logit warped time-scale
}
