% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Data2bifd.R
\name{Data2bifd}
\alias{Data2bifd}
\title{Creates a bifd object from a matrix.}
\usage{
Data2bifd(
  sargvals = NULL,
  targvals = NULL,
  y = NULL,
  sbasisobj = NULL,
  tbasisobj = NULL,
  Lfdobjs = NULL,
  Lfdobjt = NULL,
  lambdas = NULL,
  lambdat = NULL
)
}
\arguments{
\item{sargvals}{Argument values for `s' in the matrix input y(s,t).}

\item{targvals}{Argument values for `t' in the matrix input y(s,t).}

\item{y}{Data (only one) as a matrix.}

\item{sbasisobj}{basisfd object for `s'}

\item{tbasisobj}{basisfd object for `t'}

\item{Lfdobjs}{Number of derivatives used for smoothing along `s'. No smoohting will be done for NULL}

\item{Lfdobjt}{Number of derivatives used for smoothing along `t'. No smoohting will be done for NULL}

\item{lambdas}{Weight on the smoothing along `s'}

\item{lambdat}{Weight on the smoothing along `t'}
}
\description{
Creates a bifd object from a matrix.
}
