% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/eigen.fd.R
\name{eigen.fd}
\alias{eigen.fd}
\title{Takes eigen decomposition of bifd covariance object.}
\usage{
eigen.fd(cov.fd)
}
\arguments{
\item{cov.fd}{Covariance operator as a bifd object (fda package). Use \link{Data2bifd} to convert covariance matrix into a bifd object.}
}
\description{
Takes eigen decomposition of bifd covariance object.
}
