% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{Fragments}
\alias{Fragments}
\title{Spinal bone mineral density measurements (Fragments)}
\format{
Variables:
\describe{
  \item{Y_f.i}{spnbmd measurements of female subjects i=1,...,140}
  \item{X_f.i}{ages of the female subjects at time of measurements}
  \item{Y_m.i}{spnbmd measurements of male subjects i=1,...,113}
  \item{X_m.i}{ages of the male subjects at time of measurements}
}
}
\source{
\url{http://web.stanford.edu/~hastie/ElemStatLearn/}
}
\usage{
Fragments
}
\description{
Relative spinal bone mineral density (spnbmd) measurements on 261 North
American adolescents.  Each value is the difference in spnbmd
taken on two consecutive visits, divided by the average. The age is
the average age over the two visits.
}
\references{
Bachrach LK, Hastie T, Wang M-C, Narasimhan B, Marcus R. Bone Mineral 
Acquisition in Healthy Asian, Hispanic, Black and Caucasian Youth. A 
Longitudinal Study. J Clin Endocrinol Metab (1999) 84, 4702-12.
}
\keyword{datasets}
