% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{locate_crossings}
\alias{locate_crossings}
\title{Locate crossings}
\usage{
locate_crossings(x_vec, threshold, type = c("up", "down"))
}
\arguments{
\item{x_vec}{Vector which might cross the given threshold}

\item{threshold}{Threshold-vector}

\item{type}{locate 'up' or 'down' crossings?}
}
\description{
Locate crossings
}
