% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_band.R
\name{make_band_FFSCB_z}
\alias{make_band_FFSCB_z}
\title{Fast 'n' fair simultaneous confidence band (Gaussian)}
\usage{
make_band_FFSCB_z(x, diag.cov.x, tau, t0 = NULL, conf.level = 0.95,
  n_int = 10)
}
\arguments{
\item{x}{Functional parameter estimate.}

\item{diag.cov.x}{Diagonal of Cov(x), in which x is the functional estimator (for instance, the covariance function of the empirical mean function).}

\item{tau}{Pointwise standard deviation of the standardized and differentiated sample functions. Can be estimated by tau_fun().}

\item{t0}{Parameter t0 of the fast and fair simultaneous confidence bands. If left unspecified (default t0=NULL), t0 is set to the location which maximizes tau.}

\item{conf.level}{confidence level (default: 0.95)}

\item{n_int}{Number of equidistant intervals among which to allocate the type-I error rate (1-conf.level) in equal shares.}
}
\description{
Fast 'n' fair simultaneous confidence band (Gaussian)
}
\references{
Liebl, D. and Reimherr, M. (2019). Fast and fair simultaneous confidence bands.
}
