% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_band.R
\name{make_band_KR_z}
\alias{make_band_KR_z}
\title{Kac-Rice simultaneous confidence band (Gaussian)}
\usage{
make_band_KR_z(x, tau, diag.cov, N, conf.level = 0.95)
}
\arguments{
\item{x}{Functional parameter estimate.}

\item{tau}{Pointwise standard deviation of the standardized and differentiated sample functions. Can be estimated by tau_fun().}

\item{diag.cov}{The diagonal of N * Cov(X), in which X is the functional estimator.}

\item{N}{It should be '1' if 'cov' is the covariance operator for X itself, which is the default value.}

\item{conf.level}{confidence level (default: 0.95)}
}
\description{
Kac-Rice simultaneous confidence band (Gaussian)
}
