% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_cov_m}
\alias{make_cov_m}
\title{Make discretized covariance function}
\usage{
make_cov_m(cov.f = covf_st_matern, grid, cov.f.params = NULL)
}
\arguments{
\item{cov.f}{Covariance function}

\item{grid}{Evaluation-grid}

\item{cov.f.params}{Parameters of the covariance function}
}
\description{
Make discretized covariance function
}
