% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_fragm_sample}
\alias{make_fragm_sample}
\title{Make sample (for simulation)}
\usage{
make_fragm_sample(mean.v, cov.m, N, fragm_len, grid = NULL,
  dist = "rnorm", ...)
}
\arguments{
\item{mean.v}{Mean-Vector (discretized mean function)}

\item{cov.m}{Covariance-Matrix (discretized covariance function)}

\item{N}{Number of functions}

\item{fragm_len}{Number of evaluation points per fragment}

\item{grid}{Grid in [0,1]. Default: grid <- seq(0,1,length(mean.v))}

\item{dist}{Distribution}

\item{...}{further parameters for dist}
}
\description{
Make sample (for simulation)
}
