% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{make_sample}
\alias{make_sample}
\title{Make sample (for simulation)}
\usage{
make_sample(mean.v, cov.m, N, dist = "rnorm", ...)
}
\arguments{
\item{mean.v}{Mean-Vector (discretized mean function)}

\item{cov.m}{Covariance-Matrix (discretized covariance function)}

\item{N}{Number of functions}

\item{dist}{Distribution}

\item{...}{further parameters for dist}
}
\description{
Make sample (for simulation)
}
