% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf.localshift}
\alias{meanf.localshift}
\title{Meanfunction with local peak (imple shifting)}
\usage{
meanf.localshift(x, delta = 0)
}
\arguments{
\item{x}{function argument}

\item{delta}{shifting parameter}
}
\description{
Meanfunction with local peak (imple shifting)
}
