% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf.poly}
\alias{meanf.poly}
\title{Meanfunction (polynomial)}
\usage{
meanf.poly(x, params = c(0, 1))
}
\arguments{
\item{x}{function argument}

\item{params}{Parameters: params=c(shift,scale).}
}
\description{
Meanfunction (polynomial)
}
