% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf.scale}
\alias{meanf.scale}
\title{Meanfunction (polynomial simple scaling)}
\usage{
meanf.scale(x, delta = 0)
}
\arguments{
\item{x}{function argument}

\item{delta}{scaling parameter}
}
\description{
Meanfunction (polynomial simple scaling)
}
