% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf.shift}
\alias{meanf.shift}
\title{Meanfunction (polynomial simple shifting)}
\usage{
meanf.shift(x, delta = 0)
}
\arguments{
\item{x}{function argument}

\item{delta}{shifting parameter}
}
\description{
Meanfunction (polynomial simple shifting)
}
