% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf_bump}
\alias{meanf_bump}
\title{Meanfunction with local bump}
\usage{
meanf_bump(x, quarter = 1, height = 0.1)
}
\arguments{
\item{x}{function argument}

\item{quarter}{1,2,3, or 4 to select domain-quarter.}

\item{height}{height of the bump function}
}
\description{
Meanfunction with local bump
}
