% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf_peak}
\alias{meanf_peak}
\title{Meanfunction with local peak}
\usage{
meanf_peak(x, params = c(0, 1, 2, 16, 1))
}
\arguments{
\item{x}{function argument}

\item{params}{Parameters}
}
\description{
Meanfunction with local peak
}
