% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf_scale}
\alias{meanf_scale}
\title{Meanfunction (polynomial with simple scaling)}
\usage{
meanf_scale(x, delta = 0)
}
\arguments{
\item{x}{function argument}

\item{delta}{scaling parameter}
}
\description{
Meanfunction (polynomial with simple scaling)
}
\examples{
curve(meanf_scale(x, 1), from=0, to=1)
curve(meanf_scale(x, 0), from=0, to=1, lty=2, add=TRUE)
}
