% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/misc.R
\name{n_ts}
\alias{n_ts}
\title{Count number of observations in X_i(t)*X_i(s), i=1...,n, with fragmentary functional data X_i}
\usage{
n_ts(X_mat)
}
\arguments{
\item{X_mat}{nxp Matrix with fragmentary observed functional data of subjects i=1,...,p stored column-wise. Missing data are assumed to be set to NA.}
}
\description{
Count number of observations in X_i(t)*X_i(s), i=1...,n, with fragmentary functional data X_i
}
