% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.fregion.band.R
\name{plot.confidence_band}
\alias{plot.confidence_band}
\title{Visualizes confidence bands constructed from \link{confidence_band}.}
\usage{
\method{plot}{confidence_band}(band, center = TRUE,
  legendx = "topleft", legendy = NULL, ...)
}
\arguments{
\item{band}{A 'confidence_band' object, the output from \link{confidence_band} funciton.}

\item{center}{Whether to include the functional estimate or not.}

\item{legendx}{position `x' of the legend. If NULL is passed, the legend will not be drawn (However, it may be added manually)}

\item{legendy}{position `y' of the legend.}

\item{...}{Graphical parameters to be passed/overrided. If 'center' is TRUE, the first elements of 'col', 'lwd', 'lty' will be used for the estimate and the next ones will be used for the bands, but using the same values for one pair, i.e., lower and upper bounds.}
}
\description{
Visualizes confidence bands constructed from \link{confidence_band}.
}
