% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_doc.R
\docType{data}
\name{spnbmd}
\alias{spnbmd}
\title{Spinal bone mineral density measurements (spnbmd)}
\format{Variables:
\describe{
  \item{idnum}{identifies the child, and hence the repeat measurements}
  \item{age}{average age of child when measurements were taken}
  \item{gender}{male or female}
  \item{spnbmd}{Relative Spinal bone mineral density measurement}
}}
\source{
\url{http://web.stanford.edu/~hastie/ElemStatLearn/}
}
\usage{
spnbmd
}
\description{
Relative spinal bone mineral density measurements on 261 North
American adolescents.  Each value is the difference in spnbmd
taken on two consecutive visits, divided by the average. The age is
the average age over the two visits.
}
\references{
Bachrach LK, Hastie T, Wang M-C, Narasimhan B, Marcus R. Bone Mineral 
Acquisition in Healthy Asian, Hispanic, Black and Caucasian Youth. A 
Longitudinal Study. J Clin Endocrinol Metab (1999) 84, 4702-12.
}
\keyword{datasets}
