% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_fun.R
\name{tau_fragments}
\alias{tau_fragments}
\title{R-function for computing tau from fragmentary functional data.
Caution: only one single fragment per function is assumed.}
\usage{
tau_fragments(X_mat, grid_mat)
}
\arguments{
\item{X_mat}{Matrix of sample functions (nrow=p, ncol=n, p=number of discretization point, n=sample size).}

\item{grid_mat}{Matrix (pxn) of grid points over which the fragments are observed.}
}
\description{
R-function for computing tau from fragmentary functional data.
Caution: only one single fragment per function is assumed.
}
