% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tau_fun.R
\name{tau_fun}
\alias{tau_fun}
\title{Computes tau(t), the pointwise standard deviation of the standardized and differentiated sample functions}
\usage{
tau_fun(x)
}
\arguments{
\item{x}{Matrix of sample functions (nrow=p, ncol=n, p=number of discretization point, n=sample size).}
}
\value{
tau_t Pointwise standard deviation of the standardized and differentiated sample functions.
}
\description{
Computes tau(t), the pointwise standard deviation of the standardized and differentiated sample functions
}
