% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fregion.test.R
\name{testing}
\alias{testing}
\alias{fregion.test}
\title{Performs hypothesis test(s) based on functional confidence region(s).}
\usage{
fregion.test(x, x0 = 0, cov, tau = NULL, t0 = NULL, N = 1,
  type = c("FFSCB.t", "Ec"), conf.level = c(0.95), n_int = 10,
  tol = NULL, pc.cut = c(0.99), prec = NULL)
}
\arguments{
\item{x}{Functional parameter estimate. It can be either a vector or \link{fd} object from \link{fda}.}

\item{x0}{Functional parameter under the null hypothesis. Zero function is assumed if it's not given.}

\item{cov}{N * Cov(X), in which X is the functional estimator. It can be either matrix or \link{bifd} object from \link{fda}. The eigen decomposition of Cov(X) can be used instead.}

\item{tau}{Pointwise standard deviation of the standardized and differentiated sample functions. Can be estimated by tau_fun().}

\item{t0}{Parameter t0 of the fast and fair simultaneous confidence bands.}

\item{N}{It should be '1' if 'cov' is the covariance operator for X itself, which is the default value.}

\item{type}{This specifies which regions to be used for the tests. It should be a collection of the following: "FFSCB.t", "Ec".
\itemize{
  \item FFSCB.t : Test based on the fast 'n' fair simultaneous confidence band of Liebl and Reimherr (2019).
  \item Ec : Test based on the ellipsoid region in which radius are proportional to the square-root of corresponding eigenvalues as suggested in Choi and Reimherr (2018).
}}

\item{conf.level}{A vector of confidence levels for the bands to achieve.}

\item{n_int}{Number of intervals for the piecewise linear confidence bounds.}

\item{tol}{Controls the tolerance value used by stats::optimize(). The default (tol=NULL) leads to the functions' default values.}

\item{pc.cut}{It takes a vector of number of fPC to use in each HT. For integer values, fPC up to those values will be used. If it's a value from 0 to 1, this specifies the proportion of (estimated) variance that should be explained by the fPCs. If it is 0, all the available fPCs will be used as long as the size of eigenvalues are greater than .Machine$double.eps.}

\item{prec}{This determines the accuracy of \link{imhof}. One may try to modify this if p-value achieved in Ellipsoid form other than Epc gives negative value. It should the the form of c(epsabs, epsrel, limit).}
}
\description{
Performs hypothesis test(s) based on functional confidence region(s).
}
\examples{
# Generate a sample
p <- 200 ; N <- 80 ; rangeval = c(0,1)
grid  <- make.grid(p, rangevals=rangeval)
mu0   <- meanf.poly(grid,c(0,1)) ; names(mu0) = grid
mu    <- meanf.poly(grid,c(0,1.1)) ; names(mu) = grid
cov.m <- make.cov.m(cov.f = covf.st.matern, grid=grid, cov.f.params=c(2/2,1,1))
x     <- make.sample(mu,cov.m,N)

# Find the estimate and covariance
hat.mu       <- rowMeans(x)
hat.cov.m    <- crossprod(t(x - hat.mu)) / (N-1)
hat.tau.v    <- tau_fun(x)

# Compare different methods for Hypothesis testings.
(a1 <- fregion.test(x=hat.mu,x0=mu0,cov=hat.cov.m,tau=hat.tau.v,N=N,type=c("Ec"),pc.cut=c(1,3,4,5,0.99,0.999)))
}
\references{
\itemize{
   \item Liebl, D. and Reimherr, M. (2019). Fast and fair simultaneous confidence bands.
   \item Choi, H. and Reimherr, M. (2018). A geometric approach to confidence regions and bands for functional parameters. Journal of the Royal Statistical Society: Series B (Statistical Methodology) 80 239-260.
}
}
