% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_pvalue.R
\name{get_pvalue_FFSCB_t}
\alias{get_pvalue_FFSCB_t}
\title{FFSCB p-value (t-distr)}
\usage{
get_pvalue_FFSCB_t(x, x0 = NULL, tau, t0 = NULL, diag.cov, N,
  n.eval.points = 11, n_int = 5)
}
\arguments{
\item{x}{function argument}

\item{x0}{Functional parameter under the null hypothesis. Zero function is assumed if it's not given.}

\item{tau}{Pointwise standard deviation of the standardized and differentiated sample functions. Can be estimated by tau_fun().}

\item{t0}{Parameter t0 of the fast and fair simultaneous confidence bands.}

\item{diag.cov}{The diagonal of N * Cov(X), in which X is the functional estimator.}

\item{N}{It should be '1' if 'cov' is the covariance operator for X itself, which is the default value.}

\item{n.eval.points}{Number of evaluation points for the p-value function. Large values (>10) lead to slow computations.}

\item{n_int}{Number of intervals parameter used by the function make_band_FFSCB_z()}
}
\description{
FFSCB p-value (t-distr)
}
\references{
Liebl, D. and Reimherr, M. (2019). Fast and fair simultaneous confidence bands.
}
