% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf_ellipse}
\alias{meanf_ellipse}
\title{Meanfunction with local ellipse}
\usage{
meanf_ellipse(x, params = c(0, 1/4, 1/5))
}
\arguments{
\item{x}{function argument}

\item{params}{parameters params=c(start, end, height)}
}
\description{
Meanfunction with local ellipse
}
