% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf_localshift}
\alias{meanf_localshift}
\title{Meanfunction with local peak (polynomial with simple local peak)}
\usage{
meanf_localshift(x, delta = 0)
}
\arguments{
\item{x}{function argument}

\item{delta}{shifting parameter}
}
\description{
Meanfunction with local peak (polynomial with simple local peak)
}
