% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf_localshift}
\alias{meanf_localshift}
\title{Meanfunction with local peak (polynomial with simple local peak)}
\usage{
meanf_localshift(x, delta = 0)
}
\arguments{
\item{x}{function argument}

\item{delta}{shifting parameter}
}
\description{
Meanfunction with local peak (polynomial with simple local peak)
}
\examples{
curve(meanf_localshift(x, 0), from=0, to=1, lty=2)
curve(meanf_localshift(x,.1), from=0, to=1, lty=1, add=TRUE)
}
