% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf_peak}
\alias{meanf_peak}
\title{Meanfunction with local peak}
\usage{
meanf_peak(x, params = c(0, 1, 2, 16, 1))
}
\arguments{
\item{x}{function argument}

\item{params}{Parameters}
}
\description{
Meanfunction with local peak
}
\examples{
curve(meanf_peak(x, c(0,1,2,16,1)), from=0, to=1, 
main="Meanfct Peak", ylab="",xlab="")
curve(meanf_peak(x, c(-1,1,2,16,1)), from=0, to=1, 
lty=2, add=TRUE)
curve(meanf_peak(x, c(0,2,2,16,1)), from=0, to=1, 
lty=2, add=TRUE)
curve(meanf_peak(x, c(0,1,1+1,10,1)), from=0, to=1, 
lty=2, add=TRUE)
}
