% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf_poly}
\alias{meanf_poly}
\title{Meanfunction (polynomial)}
\usage{
meanf_poly(x, params = c(0, 1))
}
\arguments{
\item{x}{function argument}

\item{params}{Parameters: params=c(shift,scale).}
}
\description{
Meanfunction (polynomial)
}
\examples{
curve(meanf_poly(x, c(0,2)), from=0, to=1, 
main="Meanfct Poly", ylab="",xlab="")
curve(meanf_poly(x, c(0,1)), from=0, to=1, 
lty=2, add=TRUE)
}
