% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf_rect}
\alias{meanf_rect}
\title{Meanfunction with local rectangles (polynomial with simple local rectangles)}
\usage{
meanf_rect(x, height = 1/5, rect1 = c(0, 0.2), rect2 = c(0.8, 1))
}
\arguments{
\item{x}{function argument}

\item{height}{hight of rectangles}

\item{rect1}{location of first rectangle}

\item{rect2}{location of second rectangle}
}
\description{
Meanfunction with local rectangles (polynomial with simple local rectangles)
}
\examples{
curve(meanf_rect(x, height=1/5), from=0, to=1,
main="Meanfct Rect", ylab="",xlab="")
curve(meanf_rect(x, height=1/10), from=0, to=1, 
lty=2, add=TRUE)
}
