% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf_rect}
\alias{meanf_rect}
\title{Meanfunction with local rectangles (polynomial with simple local rectangles)}
\usage{
meanf_rect(x, delta = 1/5, rect = c(0, 0.125))
}
\arguments{
\item{x}{function argument}

\item{delta}{hight of rectangle}

\item{rect}{location of the rectangle}
}
\description{
Meanfunction with local rectangles (polynomial with simple local rectangles)
}
\examples{
curve(meanf_rect(x, delta=1/5), from=0, to=1,
main="Meanfct Rect", ylab="",xlab="")
curve(meanf_rect(x, delta=1/10), from=0, to=1, 
lty=2, add=TRUE)
}
