% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/meanf.R
\name{meanf_shift}
\alias{meanf_shift}
\title{Meanfunction (polynomial with simple shifting)}
\usage{
meanf_shift(x, delta = 0)
}
\arguments{
\item{x}{function argument}

\item{delta}{shifting parameter}
}
\description{
Meanfunction (polynomial with simple shifting)
}
\examples{
curve(meanf_shift(x, 0), from=0, to=1, lty=2)
curve(meanf_shift(x,.1), from=0, to=1, lty=1, add=TRUE)
}
